pyproject.toml
[project]
name = "autoannotator"
version = "0.1.0"
description = "Auto annotator for LeRobot-format datasets"
authors = [
    {name = "Leonardo", email = "leonardo.perez@nonhuman.site"},
]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
keywords = ["robotics", "ai", "machine-learning", "physical-intelligence"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Robotics",
]

# Primary dependencies
dependencies = [
    "datasets>=2.12.0",
    "opencv-python>=4.8.0",
    "pillow>=9.5.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    # AutoAnnotator Core
    "google-genai>=1.54.0",
    "moviepy>=2.2.1",
    "imageio-ffmpeg>=0.4.9",
    "python-dotenv>=1.1.0",
    "pydantic>=2.0.0",
    "pyarrow>=14.0.0",
    # LeRobot from GitHub
    "lerobot @ git+https://github.com/huggingface/lerobot.git",
    "huggingface-hub>=0.34.2,<0.36.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
]

# Optional dependencies
[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]
docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

# Development tools configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["autoannotator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# Package discovery
[tool.setuptools.packages.find]
exclude = ["images*", "tests*", "scripts*"]

# Build system
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
